+++ circuits/examples/bitcoin_ecdsa_threshold.rs
@@ -301,6 +301,7 @@ fn main() {
             &srs.verifier_params(),
             &vk,
             &instance,
+            None,
             &proof
         )
         .is_ok()