+++ circuits/examples/hybrid_mt.rs
@@ -374,6 +374,7 @@ fn main() {
             &srs.verifier_params(),
             &vk,
             &instance,
+            None,
             &proof
         )
         .is_ok()